app-genai/requirements.txt
# Core dependencies
pynest-api==0.3.1
asyncpg==0.29.0
sqlalchemy==2.0.19
pydantic<2.0.0
python-dotenv==1.0.0

# LangChain y procesamiento de documentos
langchain==0.1.0
langchain-community==0.0.16
chromadb==0.4.22
pypdf==3.17.1
pdfplumber==0.10.3
tiktoken==0.5.2
openai==1.12.0
ollama==0.1.6  # Añadido para Llama

# Procesamiento de datos
# Procesamiento de texto y documentos
python-docx==1.0.1  # Para generar documentos Word
httpx>=0.25.2,<0.26.0 
pandas==2.2.0

# Para manejo de archivos
python-multipart==0.0.6

# Backend y API
fastapi==0.109.2
uvicorn==0.27.1


email-validator
python-pptx==0.6.21